requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.1
idna==2.9
Pillow==7.0.0
python-slugify==4.0.0
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.9
yandex.translate==0.3.5
